requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
accelerate
openai
python-dotenv